Version 5
I improved the GUI of the movie info window, by increasing its size to the size of the movie frame in the main window, which makes it seem more integrated with the program. Also, made the side tab labels into button, they each output something different. I also added an Exit button to the program.